Implemented another fix

+++ src/main/java/core/basesyntax/FigureSupplier.java
@@ -3,9 +3,9 @@ package core.basesyntax;
 import java.util.Random;
 
 public class FigureSupplier {
-    private static final int DEFAULT_RADIUS = 10;
+    private static final double DEFAULT_RADIUS = 10;
     private static final int NUMBER_OF_FIGURES = 5;
-    private static final int MAX_UNITS = 35;
+    private static final double MAX_UNITS = 35;
     private Random random = new Random();
     private ColorSupplier colorSupplier = new ColorSupplier();
 
@@ -34,6 +34,6 @@ public class FigureSupplier {
     }
 
     public double getRandomUnits() {
-        return random.nextInt(MAX_UNITS);
+        return random.nextDouble(MAX_UNITS);
     }
 }
